apps/api/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=1.5.0
numpy>=1.24.0
python-multipart>=0.0.9
typer>=0.9.0
reportlab>=4.0.0
xlsxwriter>=3.1.0
openpyxl>=3.1.0
weasyprint>=61.0.0
celery[redis]>=5.3.6
redis>=5.0.1
slowapi>=0.1.9
pdfplumber>=0.11.0
openai>=1.0.0
pydantic-settings>=2.2.1
